refactor: генератор id перенесён в другой пакет
